requirements.txt
click
dateutil
py-trello
